// frontend/src/api/model/spectralRecordArtifact.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { ArtifactTypeEnum } from './artifactTypeEnum';
import type { File } from './file';

/**
 * Serializer for spectral record artifacts.
 */
export interface SpectralRecordArtifact {
  readonly id: string;
  /** Type of artifact (e.g. histogram, processed spectral logs, ...)

* `spectral` - Processed log file into spectral file (Parquet) */
  artifact_type: ArtifactTypeEnum;
  readonly artifact: File;
  /** Reference to SpectralRecord to which this artifact belongs */
  spectral_record: string;
  /** Time when the artifact was created */
  readonly created_at: string;
}